# document/swagger.yaml
openapi: 3.0.0
info:
  title: Charum
  description: |-
    This is a Charum Back End
  version: 1.0.0
servers:
  - url: https://charum-api.nyakit.in/api/v1
...
paths:
...
  /admin/topic/{id}:
    get:
      tags:
        - Admin
      summary: Get topic data
      description: Get topic data by id
      operationId: GetTopicByID
      parameters:
        - in: path
          name: id
          description: topic ID
          required: true
          schema:
            type: string
            example: 637dfc9bb09d0bdd7a2fe3de
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to get topic by id
                  data:
                    type: object
                    properties:
                      topic:
                        $ref: '#/components/schemas/Topic'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: failed to get topic
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null

    put:
      tags:
        - Admin
      summary: Update topic data
      description: Update topic data
      operationId: UpdateTopic
      parameters:
        - in: path
          name: id
          description: topic ID
          required: true
          schema:
            type: string
            example: 637dfc9bb09d0bdd7a2fe3de
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                topic:
                  type: string
                  example: new Topic
                description:
                  type: string
                  example: new Description
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success update topic
                  data:
                    type: object
                    properties:
                      topic:
                        $ref: '#/components/schemas/Topic'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: topic already exist
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: failed to update topic
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
    delete:
      tags:
        - Admin
      summary: Delete topic
      description: Delete topic
      operationId: DeleteTopic
      parameters:
        - in: path
          name: id
          description: topic ID
          required: true
          schema:
            type: string
            example: 637dfc9bb09d0bdd7a2fe3de
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to delete topic
                  data:
                    type: object
                    properties:
                      topic:
                        $ref: '#/components/schemas/Topic'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: unauthorized
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: failed to delete topic
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Topic:
      type: object
      properties:
        _id:
          type: string
          example: 6379fe8dad33ec4251700ded
        topic:
          type: string
          example: topic name
        description:
          type: string
          example: topic description
        createdAt:
          type: string
          example: 2022-11-26T23:05:11.806+07:00
        updatedAt:
          type: string
          example: 2022-11-26T23:05:11.806+07:00
...
security:
  - bearerAuth: []